# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getVersion","summary":"Check version of the application","parameters":[],"responses":{"200":{"description":"Retrieve version of API"}},"tags":["Version"]}},"/health":{"get":{"operationId":"HealthController_check","summary":"Check health of the application","parameters":[],"responses":{"200":{"description":"API Health Check"},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}},"tags":["Health"]}},"/user":{"get":{"operationId":"UserController_getAllUsers","summary":"Retrieve all users","parameters":[{"name":"Authorization","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"Retrieve all users"}},"tags":["User"]},"post":{"operationId":"UserController_createUser","summary":"Create a single user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/user/{id}":{"get":{"operationId":"UserController_getUser","summary":"Get single user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]},"delete":{"operationId":"UserController_deleteUser","summary":"Delete single user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]},"put":{"operationId":"UserController_updateUser","summary":"Update single user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"200":{"description":""}},"tags":["User"]}},"/user/deactivate/{id}":{"put":{"operationId":"UserController_deactivateUser","summary":"Deactivate user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/user/reset-password/{id}":{"put":{"operationId":"UserController_resetUserPassword","summary":"Send reset password email","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}}},"info":{"title":"Nest JS Starter","description":"The application for Nest JS","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UserDTO":{"type":"object","properties":{"email":{"type":"string","description":"Email address of user"},"password":{"type":"string","description":"Password of user"},"firstName":{"type":"string","description":"First name of user"},"lastName":{"type":"string","description":"Last name of user"}},"required":["email","password","firstName","lastName"]}}}}